move and tidy rabbit client
